requirements.txt
google-cloud-storage
google-api-python-client
pytest